backend/team12/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
coverage==6.5.0
Django
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.14.0
pytz==2022.4
screen==1.0.1
sqlparse==0.4.3
xlwt==1.3.0
pylint==2.15.5
factory-boy==3.2.1
djangorestframework-simplejwt==5.2.2
black==22.12.0